+++ Beam/extensions/beam-module/database/schema/mysql-schema.sql
@@ -9,12 +9,12 @@ DROP TABLE IF EXISTS `accounts`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `accounts` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `uuid` char(36) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `uuid` char(36) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
-  KEY `accounts_uuid_index` (`uuid`)
+  UNIQUE KEY `accounts_uuid_unique` (`uuid`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 /*!40101 SET character_set_client = @saved_cs_client */;
 DROP TABLE IF EXISTS `article_aggregated_views`;
@@ -23,8 +23,8 @@ DROP TABLE IF EXISTS `article_aggregated_views`;
 CREATE TABLE `article_aggregated_views` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
   `article_id` int unsigned NOT NULL,
-  `user_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `browser_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `user_id` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `browser_id` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `date` date NOT NULL,
   `pageviews` int NOT NULL DEFAULT '0',
   `timespent` int NOT NULL DEFAULT '0',
@@ -56,8 +56,8 @@ DROP TABLE IF EXISTS `article_meta`;
 CREATE TABLE `article_meta` (
   `id` bigint unsigned NOT NULL AUTO_INCREMENT,
   `article_id` int unsigned NOT NULL,
-  `key` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `value` text CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `key` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `value` text COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -136,8 +136,8 @@ DROP TABLE IF EXISTS `article_titles`;
 CREATE TABLE `article_titles` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
   `article_id` int unsigned NOT NULL,
-  `variant` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `title` varchar(768) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `variant` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `title` varchar(768) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -151,9 +151,9 @@ DROP TABLE IF EXISTS `article_views_snapshots`;
 CREATE TABLE `article_views_snapshots` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
   `time` timestamp NOT NULL,
-  `property_token` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `external_article_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `referer_medium` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `property_token` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `external_article_id` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `referer_medium` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `count` int unsigned NOT NULL,
   PRIMARY KEY (`id`),
   KEY `article_views_snapshots_time_index` (`time`),
@@ -168,12 +168,12 @@ DROP TABLE IF EXISTS `articles`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `articles` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `external_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `property_uuid` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `title` varchar(768) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `url` varchar(768) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `content_type` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `image_url` varchar(768) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `external_id` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `property_uuid` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `title` varchar(768) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `url` varchar(768) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `content_type` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `image_url` varchar(768) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `published_at` timestamp NULL DEFAULT NULL,
   `pageviews_all` bigint NOT NULL DEFAULT '0',
   `pageviews_signed_in` bigint NOT NULL DEFAULT '0',
@@ -196,8 +196,8 @@ DROP TABLE IF EXISTS `authors`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `authors` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `external_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `external_id` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -209,8 +209,8 @@ DROP TABLE IF EXISTS `config_categories`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `config_categories` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `code` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `display_name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `code` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `display_name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -222,13 +222,13 @@ DROP TABLE IF EXISTS `configs`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `configs` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `config_category_id` int unsigned NOT NULL,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `display_name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `config_category_id` int unsigned DEFAULT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `display_name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `property_id` int unsigned DEFAULT NULL,
-  `value` text CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci,
-  `description` text CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci,
-  `type` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL DEFAULT 'text',
+  `value` text COLLATE utf8mb4_unicode_ci,
+  `description` text COLLATE utf8mb4_unicode_ci,
+  `type` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL DEFAULT 'text',
   `sorting` int NOT NULL DEFAULT '10',
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
@@ -246,7 +246,7 @@ DROP TABLE IF EXISTS `conversion_commerce_event_products`;
 CREATE TABLE `conversion_commerce_event_products` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
   `conversion_commerce_event_id` int unsigned NOT NULL,
-  `product_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `product_id` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   PRIMARY KEY (`id`),
   KEY `event_foreign` (`conversion_commerce_event_id`),
   CONSTRAINT `event_foreign` FOREIGN KEY (`conversion_commerce_event_id`) REFERENCES `conversion_commerce_events` (`id`) ON DELETE CASCADE
@@ -261,14 +261,14 @@ CREATE TABLE `conversion_commerce_events` (
   `time` timestamp NOT NULL,
   `minutes_to_conversion` int NOT NULL,
   `event_prior_conversion` int unsigned NOT NULL,
-  `step` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `funnel_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `amount` double(8,2) DEFAULT NULL,
-  `currency` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_campaign` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_content` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_medium` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_source` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `step` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `funnel_id` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `amount` double DEFAULT NULL,
+  `currency` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_campaign` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_content` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_medium` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_source` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -285,12 +285,12 @@ CREATE TABLE `conversion_general_events` (
   `time` timestamp NOT NULL,
   `minutes_to_conversion` int NOT NULL,
   `event_prior_conversion` int unsigned NOT NULL,
-  `action` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `category` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_campaign` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_content` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_medium` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_source` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `action` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `category` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_campaign` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_content` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_medium` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_source` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -311,10 +311,10 @@ CREATE TABLE `conversion_pageview_events` (
   `locked` tinyint(1) DEFAULT NULL,
   `signed_in` tinyint(1) DEFAULT NULL,
   `timespent` int unsigned DEFAULT NULL,
-  `rtm_campaign` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_content` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_medium` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `rtm_source` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_campaign` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_content` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_medium` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `rtm_source` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -330,10 +330,10 @@ DROP TABLE IF EXISTS `conversion_sources`;
 CREATE TABLE `conversion_sources` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
   `conversion_id` int unsigned NOT NULL,
-  `type` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `referer_medium` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `referer_source` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `referer_host_with_path` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `type` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `referer_medium` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `referer_source` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `referer_host_with_path` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `article_id` int unsigned DEFAULT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
@@ -350,11 +350,11 @@ DROP TABLE IF EXISTS `conversions`;
 CREATE TABLE `conversions` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
   `article_id` int unsigned NOT NULL,
-  `user_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `amount` double(8,2) NOT NULL,
-  `currency` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `user_id` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `amount` double NOT NULL,
+  `currency` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `paid_at` timestamp NOT NULL,
-  `transaction_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `transaction_id` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `events_aggregated` tinyint(1) NOT NULL DEFAULT '0',
   `source_processed` tinyint(1) NOT NULL DEFAULT '0',
   `created_at` timestamp NULL DEFAULT NULL,
@@ -389,7 +389,7 @@ DROP TABLE IF EXISTS `entities`;
 CREATE TABLE `entities` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
   `parent_id` int unsigned DEFAULT NULL,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   `deleted_at` timestamp NULL DEFAULT NULL,
@@ -404,8 +404,8 @@ CREATE TABLE `entity_params` (
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   `entity_id` int unsigned NOT NULL,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `type` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `type` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `deleted_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
   KEY `entity_params_entity_id_foreign` (`entity_id`),
@@ -417,11 +417,11 @@ DROP TABLE IF EXISTS `failed_jobs`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `failed_jobs` (
   `id` bigint unsigned NOT NULL AUTO_INCREMENT,
-  `uuid` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `connection` text CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `queue` text CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `payload` longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `exception` longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `uuid` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `connection` text COLLATE utf8mb4_unicode_ci NOT NULL,
+  `queue` text COLLATE utf8mb4_unicode_ci NOT NULL,
+  `payload` longtext COLLATE utf8mb4_unicode_ci NOT NULL,
+  `exception` longtext COLLATE utf8mb4_unicode_ci NOT NULL,
   `failed_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (`id`),
   UNIQUE KEY `failed_jobs_uuid_unique` (`uuid`)
@@ -432,8 +432,8 @@ DROP TABLE IF EXISTS `jobs`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `jobs` (
   `id` bigint unsigned NOT NULL AUTO_INCREMENT,
-  `queue` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `payload` longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `queue` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `payload` longtext COLLATE utf8mb4_unicode_ci NOT NULL,
   `attempts` tinyint unsigned NOT NULL,
   `reserved_at` int unsigned DEFAULT NULL,
   `available_at` int unsigned NOT NULL,
@@ -447,7 +447,7 @@ DROP TABLE IF EXISTS `migrations`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `migrations` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `migration` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `migration` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `batch` int NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
@@ -457,18 +457,18 @@ DROP TABLE IF EXISTS `newsletters`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `newsletters` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `mailer_generator_id` int unsigned NOT NULL,
-  `segment` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `mail_type_code` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `criteria` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `segment` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `mail_type_code` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `criteria` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `articles_count` int unsigned NOT NULL,
-  `recurrence_rule` text CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci,
+  `recurrence_rule` text COLLATE utf8mb4_unicode_ci,
   `personalized_content` tinyint(1) NOT NULL,
-  `timespan` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `state` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `email_subject` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `email_from` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `timespan` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `state` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `email_subject` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `email_from` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `last_sent_at` timestamp NULL DEFAULT NULL,
   `starts_at` timestamp NULL DEFAULT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
@@ -481,14 +481,14 @@ DROP TABLE IF EXISTS `properties`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `properties` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `uuid` char(36) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `uuid` char(36) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `account_id` int unsigned NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
+  UNIQUE KEY `properties_uuid_unique` (`uuid`),
   KEY `properties_account_id_foreign` (`account_id`),
-  KEY `properties_uuid_index` (`uuid`),
   CONSTRAINT `properties_account_id_foreign` FOREIGN KEY (`account_id`) REFERENCES `accounts` (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 /*!40101 SET character_set_client = @saved_cs_client */;
@@ -497,8 +497,8 @@ DROP TABLE IF EXISTS `referer_medium_labels`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `referer_medium_labels` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `referer_medium` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `label` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `referer_medium` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `label` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -510,8 +510,8 @@ DROP TABLE IF EXISTS `sections`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `sections` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `external_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `external_id` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -524,7 +524,7 @@ DROP TABLE IF EXISTS `segment_browsers`;
 CREATE TABLE `segment_browsers` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
   `segment_id` int unsigned NOT NULL,
-  `browser_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `browser_id` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -537,9 +537,9 @@ DROP TABLE IF EXISTS `segment_groups`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `segment_groups` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `code` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `type` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `code` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `type` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `sorting` int NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
@@ -552,15 +552,15 @@ DROP TABLE IF EXISTS `segment_rules`;
 CREATE TABLE `segment_rules` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
   `segment_id` int unsigned NOT NULL,
-  `event_category` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `event_action` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `event_category` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `event_action` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `timespan` int DEFAULT NULL,
   `count` int NOT NULL,
-  `fields` longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL COMMENT 'JSON encoded fields',
+  `fields` longtext COLLATE utf8mb4_unicode_ci NOT NULL COMMENT 'JSON encoded fields',
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
-  `operator` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `flags` longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL COMMENT 'JSON encoded flags',
+  `operator` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `flags` longtext COLLATE utf8mb4_unicode_ci NOT NULL COMMENT 'JSON encoded flags',
   PRIMARY KEY (`id`),
   KEY `segment_rules_segment_id_foreign` (`segment_id`),
   CONSTRAINT `segment_rules_segment_id_foreign` FOREIGN KEY (`segment_id`) REFERENCES `segments` (`id`)
@@ -572,7 +572,7 @@ DROP TABLE IF EXISTS `segment_users`;
 CREATE TABLE `segment_users` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
   `segment_id` int unsigned NOT NULL,
-  `user_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `user_id` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -585,8 +585,8 @@ DROP TABLE IF EXISTS `segments`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `segments` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `code` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `code` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `active` tinyint(1) NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
@@ -604,8 +604,8 @@ DROP TABLE IF EXISTS `tag_categories`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `tag_categories` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
-  `external_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
+  `external_id` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`)
@@ -630,8 +630,8 @@ DROP TABLE IF EXISTS `tags`;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `tags` (
   `id` int unsigned NOT NULL AUTO_INCREMENT,
-  `external_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `name` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `external_id` varchar(191) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
+  `name` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
@@ -642,7 +642,7 @@ DROP TABLE IF EXISTS `views_per_browser_mv`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `views_per_browser_mv` (
-  `browser_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `browser_id` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `total_views_last_30_days` int unsigned NOT NULL DEFAULT '0',
   `total_views_last_60_days` int unsigned NOT NULL DEFAULT '0',
   `total_views_last_90_days` int unsigned NOT NULL DEFAULT '0',
@@ -653,7 +653,7 @@ DROP TABLE IF EXISTS `views_per_user_mv`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `views_per_user_mv` (
-  `user_id` varchar(191) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
+  `user_id` varchar(191) COLLATE utf8mb4_unicode_ci NOT NULL,
   `total_views_last_30_days` int unsigned NOT NULL DEFAULT '0',
   `total_views_last_60_days` int unsigned NOT NULL DEFAULT '0',
   `total_views_last_90_days` int unsigned NOT NULL DEFAULT '0',
@@ -709,43 +709,41 @@ INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (39,'2018_12_10_123
 INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (40,'2018_12_19_093646_add_events_aggregated_to_conversions_table',1);
 INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (41,'2019_01_03_142538_create_article_titles',1);
 INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (42,'2019_01_24_152049_segments_add_criteria',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (43,'2019_01_29_142418_seed_config_values',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (44,'2019_01_31_142723_create_article_aggregated_views_with_id_table',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (45,'2019_02_06_144306_lock_author_segments_configuration',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (46,'2019_02_20_113557_alter_newsletters_change_timespan_type',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (47,'2019_06_13_111140_create_article_views_snapshots_table',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (48,'2019_07_02_082320_removing_paid_at_defaults',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (49,'2019_08_15_124622_seed_config_values2',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (50,'2019_09_05_085126_make_configs_property_specific',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (51,'2019_09_05_093404_create_config_categories_table',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (52,'2019_09_05_223346_seed_config_values3',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (53,'2019_10_03_181106_make_title_nullable_in_article_titles_table',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (54,'2019_11_07_153448_add_indexes_to_article_views_snapshots',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (55,'2019_12_03_123733_create_referer_medium_labels_table',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (56,'2020_01_31_140157_article_tags',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (57,'2020_02_18_201353_remove_unused_columns_in_article_views_snapshots',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (58,'2020_03_18_122311_add_external_id_column_to_sections_tags_authors_tables',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (59,'2020_04_06_120804_add_indexes_to_segments',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (60,'2020_07_28_091554_create_conversion_sources_table',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (61,'2020_08_07_114042_remove_locked_column_in_configs',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (62,'2020_08_07_122341_seed_author_segments_config_category',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (63,'2020_08_11_082125_alter_configs_remove_nullable_config_category',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (64,'2020_08_27_111750_add_content_type_to_articles',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (65,'2020_09_18_114850_alter_segment_rules_pageview_load_flags',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (66,'2020_10_09_131803_alter_conversions_remove_on_update_from_paid_at_column',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (67,'2020_11_09_081647_create_sections_segment_group',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (68,'2020_11_13_093901_seed_section_segment_config_category_and_values',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (69,'2020_12_02_090920_change_authors_segment_days_threshold_description',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (70,'2020_12_04_145757_convert_utm_to_rtm',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (71,'2020_12_16_102140_article_content_type_index',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (72,'2021_05_06_135333_create_tag_categories_table',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (73,'2021_05_06_142318_create_tag_tag_categories_table',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (74,'2021_05_11_113758_article_pageviews_indexes',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (75,'2021_05_31_000000_add_uuid_to_failed_jobs_table',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (76,'2021_06_10_141729_purge_invalid_aggregated_data',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (77,'2021_11_08_084315_add_index_to_conversions_paid_at_column',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (78,'2022_02_25_130852_dashboard_articles',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (79,'2023_04_04_070759_article_views_snapshots_add_index_time_property_token',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (80,'2023_09_07_080805_create_article_meta_table',1);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (81,'2024_04_15_131551_remove_configs_autoload_flag',2);
-INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (82,'2024_07_29_045343_drop_pageview_devices_and_referers',2);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (43,'2019_01_31_142723_create_article_aggregated_views_with_id_table',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (44,'2019_02_06_144306_lock_author_segments_configuration',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (45,'2019_02_20_113557_alter_newsletters_change_timespan_type',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (46,'2019_06_13_111140_create_article_views_snapshots_table',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (47,'2019_07_02_082320_removing_paid_at_defaults',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (48,'2019_09_05_085126_make_configs_property_specific',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (49,'2019_09_05_093404_create_config_categories_table',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (50,'2019_09_05_223346_seed_config_values3',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (51,'2019_10_03_181106_make_title_nullable_in_article_titles_table',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (52,'2019_11_07_153448_add_indexes_to_article_views_snapshots',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (53,'2019_12_03_123733_create_referer_medium_labels_table',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (54,'2020_01_31_140157_article_tags',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (55,'2020_02_18_201353_remove_unused_columns_in_article_views_snapshots',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (56,'2020_03_18_122311_add_external_id_column_to_sections_tags_authors_tables',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (57,'2020_04_06_120804_add_indexes_to_segments',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (58,'2020_07_28_091554_create_conversion_sources_table',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (59,'2020_08_07_114042_remove_locked_column_in_configs',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (60,'2020_08_07_122341_seed_author_segments_config_category',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (61,'2020_08_27_111750_add_content_type_to_articles',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (62,'2020_09_18_114850_alter_segment_rules_pageview_load_flags',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (63,'2020_10_09_131803_alter_conversions_remove_on_update_from_paid_at_column',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (64,'2020_11_09_081647_create_sections_segment_group',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (65,'2020_11_13_093901_seed_section_segment_config_category_and_values',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (66,'2020_12_02_090920_change_authors_segment_days_threshold_description',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (67,'2020_12_04_145757_convert_utm_to_rtm',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (68,'2020_12_16_102140_article_content_type_index',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (69,'2021_05_06_135333_create_tag_categories_table',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (70,'2021_05_06_142318_create_tag_tag_categories_table',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (71,'2021_05_11_113758_article_pageviews_indexes',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (72,'2021_05_31_000000_add_uuid_to_failed_jobs_table',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (73,'2021_06_10_141729_purge_invalid_aggregated_data',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (74,'2021_11_08_084315_add_index_to_conversions_paid_at_column',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (75,'2022_02_25_130852_dashboard_articles',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (76,'2023_04_04_070759_article_views_snapshots_add_index_time_property_token',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (77,'2023_09_07_080805_create_article_meta_table',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (78,'2024_04_15_131551_remove_configs_autoload_flag',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (79,'2024_07_29_045343_drop_pageview_devices_and_referers',1);
+INSERT INTO `migrations` (`id`, `migration`, `batch`) VALUES (80,'2026_06_05_000000_properties_uuid_unique_index',1);